Change Drag Proxy Glyph sizes to doubles following change to allow use of Group as a drag proxy.

+++ wires-bayesian-network/wires-bayesian-network-client/src/main/java/com/bayesian/network/client/factory/BayesianNodePaletteFactory.java
@@ -68,12 +68,12 @@ public class BayesianNodePaletteFactory extends AbstractBaseFactory<Rectangle> {
     }
 
     @Override
-    protected int getWidth() {
+    protected double getWidth() {
         return SHAPE_SIZE_X;
     }
 
     @Override
-    protected int getHeight() {
+    protected double getHeight() {
         return SHAPE_SIZE_Y;
     }
